move LinkText & update imports
