+++ src/app/Components/Bidding/Screens/ConfirmBid.tests.tsx
@@ -13,15 +13,13 @@ import { __globalStoreTestUtils__ } from "app/store/GlobalStore"
 import { renderWithWrappers } from "app/tests/renderWithWrappers"
 import { waitUntil } from "app/tests/waitUntil"
 import { merge } from "lodash"
-import { Button, Sans, Serif, Text } from "palette"
-import { Checkbox } from "palette/elements/Checkbox"
+import { Button, Checkbox, LinkText, Sans, Serif, Text } from "palette"
 import React from "react"
 import "react-native"
 import { TouchableWithoutFeedback } from "react-native"
 import relay from "react-relay"
 // @ts-expect-error STRICTNESS_MIGRATION --- 🚨 Unsafe legacy code 🚨 Please delete this and fix any type errors if you have time 🙏
 import stripe from "tipsi-stripe"
-import { LinkText } from "../../Text/LinkText"
 import { BidInfoRow } from "../Components/BidInfoRow"
 import { Address } from "../types"
 import { BidResultScreen } from "./BidResult"
