+++ src/app/Components/Bidding/Screens/ConfirmBid/index.tsx
@@ -18,7 +18,6 @@ import { PriceSummary } from "app/Components/Bidding/Screens/ConfirmBid/PriceSum
 import { Address, Bid, PaymentCardTextFieldParams, StripeToken } from "app/Components/Bidding/types"
 import { FancyModalHeader } from "app/Components/FancyModal/FancyModalHeader"
 import { Modal } from "app/Components/Modal"
-import { LinkText } from "app/Components/Text/LinkText"
 import { LegacyNativeModules } from "app/NativeModules/LegacyNativeModules"
 import { navigate } from "app/navigation/navigate"
 import { partnerName } from "app/Scenes/Artwork/Components/ArtworkExtraLinks/partnerName"
@@ -26,8 +25,7 @@ import { unsafe_getFeatureFlag } from "app/store/GlobalStore"
 import NavigatorIOS from "app/utils/__legacy_do_not_use__navigator-ios-shim"
 import { Schema, screenTrack, track } from "app/utils/track"
 import { get, isEmpty } from "lodash"
-import { Box, Button, Serif, Text, Theme } from "palette"
-import { Checkbox } from "palette/elements/Checkbox"
+import { Box, Button, Checkbox, LinkText, Serif, Text, Theme } from "palette"
 import React from "react"
 import { Image, ScrollView, ViewProps } from "react-native"
 import { commitMutation, createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
