+++ src/app/Components/Markdown.tests.tsx
@@ -2,8 +2,8 @@ import { renderWithWrappers } from "app/tests/renderWithWrappers"
 import React from "react"
 import { Linking, Text } from "react-native"
 
+import { LinkText } from "palette"
 import { Markdown } from "./Markdown"
-import { LinkText } from "./Text/LinkText"
 
 import { navigate } from "app/navigation/navigate"
 import { defaultRules } from "app/utils/renderMarkdown"
