+++ src/app/Scenes/AuctionResultsForArtistsYouFollow/AuctionResultsForArtistsYouFollow.tsx
@@ -5,7 +5,6 @@ import { ArtworkFiltersStoreProvider } from "app/Components/ArtworkFilter/Artwor
 import { PAGE_SIZE } from "app/Components/constants"
 import { AuctionResultListItemFragmentContainer } from "app/Components/Lists/AuctionResultListItem"
 import { PageWithSimpleHeader } from "app/Components/PageWithSimpleHeader"
-import { LinkText } from "app/Components/Text/LinkText"
 import { navigate } from "app/navigation/navigate"
 import { defaultEnvironment } from "app/relay/createEnvironment"
 import { extractNodes } from "app/utils/extractNodes"
@@ -16,7 +15,7 @@ import { screen } from "app/utils/track/helpers"
 import { useScreenDimensions } from "app/utils/useScreenDimensions"
 import { groupBy } from "lodash"
 import moment from "moment"
-import { Flex, Separator, Spacer, Spinner, Text } from "palette"
+import { Flex, LinkText, Separator, Spacer, Spinner, Text } from "palette"
 import React, { useState } from "react"
 import { RefreshControl, SectionList } from "react-native"
 import { RelayPaginationProp } from "react-relay"
