+++ src/app/Scenes/Fair/FairMoreInfo.tests.tsx
@@ -1,8 +1,7 @@
 import { FairMoreInfoTestsQuery } from "__generated__/FairMoreInfoTestsQuery.graphql"
 import { LocationMapContainer } from "app/Components/LocationMap/LocationMap"
-import { LinkText } from "app/Components/Text/LinkText"
 import { renderWithWrappers } from "app/tests/renderWithWrappers"
-import { Text } from "palette"
+import { LinkText, Text } from "palette"
 import React from "react"
 import { graphql, QueryRenderer } from "react-relay"
 import { ReactTestRenderer } from "react-test-renderer"
