+++ src/app/Scenes/Inbox/Components/Conversations/OrderUpdate.tests.tsx
@@ -1,11 +1,10 @@
 import React from "react"
 
 import { OrderUpdateTestsQuery } from "__generated__/OrderUpdateTestsQuery.graphql"
-import { LinkText } from "app/Components/Text/LinkText"
 import { navigate } from "app/navigation/navigate"
 import { extractText } from "app/tests/extractText"
 import { renderWithWrappers } from "app/tests/renderWithWrappers"
-import { AlertCircleFillIcon, MoneyFillIcon } from "palette"
+import { AlertCircleFillIcon, LinkText, MoneyFillIcon } from "palette"
 import "react-native"
 import { QueryRenderer } from "react-relay"
 import { act } from "react-test-renderer"
