+++ src/app/Scenes/PrivacyRequest/PrivacyRequest.tests.tsx
@@ -1,9 +1,8 @@
-import { LinkText } from "app/Components/Text/LinkText"
 import { LegacyNativeModules } from "app/NativeModules/LegacyNativeModules"
 import { navigate } from "app/navigation/navigate"
 // @ts-expect-error STRICTNESS_MIGRATION --- 🚨 Unsafe legacy code 🚨 Please delete this and fix any type errors if you have time 🙏
 import { shallow } from "enzyme"
-import { Button } from "palette"
+import { Button, LinkText } from "palette"
 import React from "react"
 import { PrivacyRequest } from "./PrivacyRequest"
