+++ src/app/utils/renderMarkdown.tsx
@@ -1,8 +1,7 @@
-import { LinkText } from "app/Components/Text/LinkText"
 import { navigate } from "app/navigation/navigate"
 import { decode } from "html-entities"
 import _ from "lodash"
-import { ClassTheme, Sans, Separator, Serif, Text } from "palette"
+import { ClassTheme, LinkText, Sans, Separator, Serif, Text } from "palette"
 import React from "react"
 import { Text as RNText, View } from "react-native"
 import SimpleMarkdown, { ParserRule, ParserRules, ReactNodeOutput } from "simple-markdown"
