+++ src/palette/elements/Checkbox/Checkbox.stories.tsx
@@ -1,6 +1,5 @@
 import { storiesOf } from "@storybook/react-native"
-import { LinkText } from "app/Components/Text/LinkText"
-import { Text } from "palette"
+import { LinkText, Text } from "palette"
 import React from "react"
 import { withTheme } from "storybook/decorators"
 import { List } from "storybook/helpers"
