+++ src/palette/elements/Input/PhoneInput/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PhoneInput"
