+++ src/palette/elements/Text/index.tsx
@@ -5,6 +5,8 @@ export * from "./helpers"
 
 export * from "./Text"
 
+export * from "./LinkText"
+
 import { Text, TextProps } from "./Text"
 
 // TextV1
